Add AbstractLabel and AbstractImageLabel to fix raw types
